+++ lib/tasks/buildThemes.js
@@ -5,8 +5,12 @@ const log = getLogger("builder:tasks:buildThemes");
 import {fileURLToPath} from "node:url";
 import os from "node:os";
 import workerpool from "workerpool";
-import {createResource} from "@ui5/fs/resourceFactory";
-import {performance} from "node:perf_hooks";
+import {createResource, createAdapter} from "@ui5/fs/resourceFactory";
+import fs from "graceful-fs";
+import {promisify} from "node:util";
+const mkdtemp = promisify(fs.mkdtemp);
+const mkdir = promisify(fs.mkdir);
+import {rimraf} from "rimraf";
 
 let pool;
 
@@ -70,8 +74,6 @@ export default async function({
 		projectName, inputPattern, librariesPattern, themesPattern, compress, cssVariables, useWorkers = false
 	}
 }) {
-	const timer = performance.now();
-
 	const combo = new ReaderCollectionPrioritized({
 		name: `theme - prioritize workspace over dependencies: ${projectName}`,
 		readers: [workspace, dependencies]
@@ -182,26 +184,14 @@ export default async function({
 	}
 
 	const allResources = await combo.byGlob("/resources/**/*.less");
-	const makeTransferableResource = async (resourceCollection) =>
-		Promise.all(
-			resourceCollection.map(async (res) => {
-				return {
-					string: await res.getString(),
-					name: res.getName(),
-					path: res.getPath(),
-				};
-			})
-		);
-
-
-	const allResourcesTransferable = await makeTransferableResource(allResources);
+	const tempDir = await createTmpDir(projectName);
+	await addResourcesToCollection(allResources, tempDir);
 
 	await Promise.all(themeResources.map(async (themeRes) => {
 		const processedResources = await buildTheme({
-			allResources: allResourcesTransferable,
-			themeResources: await makeTransferableResource([themeRes]),
+			themeResources: [themeRes.getPath()],
 			options: {
-				projectName,
+				tempDir,
 				compress,
 				cssVariables: !!cssVariables,
 			},
@@ -212,4 +202,29 @@ export default async function({
 			return workspace.write(createResource(rawResource));
 		}));
 	}));
+
+	// Cleanup
+	await rimraf(tempDir);
+}
+
+async function createTmpDir(projectName) {
+	const sanitizedProjectName = projectName.replace(/[^A-Za-z0-9]/g, "");
+
+	const tmpRootPath = path.join(os.tmpdir(), "ui5-tooling");
+	await mkdir(tmpRootPath, {recursive: true});
+
+	// Appending minus sign also because node docs advise to "avoid trailing X characters in prefix"
+	return mkdtemp(path.join(tmpRootPath, `theme-build-${sanitizedProjectName}-`));
+}
+
+
+async function addResourcesToCollection(resources, targetPath) {
+	const fsTarget = createAdapter({
+		fsBasePath: targetPath,
+		virBasePath: "/resources/"
+	});
+
+	// write all resources to the tmp folder
+	await Promise.all(resources.map((resource) => fsTarget.write(resource)));
+	return fsTarget;
 }
